[bicycle routing] Moving some data stuctures from turn_generator unit to separate files.
