+++ routing/turns_generator.hpp
@@ -1,10 +1,13 @@
 #pragma once
 
+#include "routing/loaded_path_segment.hpp"
 #include "routing/osrm2feature_map.hpp"
 #include "routing/osrm_engine.hpp"
+#include "routing/routing_result_graph.hpp"
 #include "routing/route.hpp"
 #include "routing/router.hpp"
 #include "routing/turns.hpp"
+#include "routing/turn_candidate.hpp"
 
 #include "std/function.hpp"
 #include "std/utility.hpp"
@@ -28,88 +31,6 @@ namespace turns
  * \brief Returns a segment index by STL-like range [s, e) of segments indices for the passed node.
  */
 using TGetIndexFunction = function<size_t(pair<size_t, size_t>)>;
-
-/*!
- * \brief The LoadedPathSegment struct is a representation of a single osrm node path.
- * It unpacks and stores information about path and road type flags.
- * Postprocessing must read information from the structure and does not initiate disk readings.
- */
-struct LoadedPathSegment
-{
-  vector<m2::PointD> m_path;
-  ftypes::HighwayClass m_highwayClass;
-  bool m_onRoundabout;
-  bool m_isLink;
-  EdgeWeight m_weight;
-  string m_name;
-  NodeID m_nodeId;
-  vector<SingleLaneInfo> m_lanes;
-
-  // General constructor.
-  LoadedPathSegment(RoutingMapping & mapping, Index const & index,
-                    RawPathData const & osrmPathSegment);
-  // Special constructor for side nodes. Splits OSRM node by information from the FeatureGraphNode.
-  LoadedPathSegment(RoutingMapping & mapping, Index const & index,
-                    RawPathData const & osrmPathSegment, FeatureGraphNode const & startGraphNode,
-                    FeatureGraphNode const & endGraphNode, bool isStartNode, bool isEndNode);
-
-private:
-  // Load information about road, that described as the sequence of FtSegs and start/end indexes in
-  // in it. For the side case, it has information about start/end graph nodes.
-  void LoadPathGeometry(buffer_vector<OsrmMappingTypes::FtSeg, 8> const & buffer, size_t startIndex,
-                        size_t endIndex, Index const & index, RoutingMapping & mapping,
-                        FeatureGraphNode const & startGraphNode,
-                        FeatureGraphNode const & endGraphNode, bool isStartNode, bool isEndNode);
-};
-
-/*!
- * \brief The TurnCandidate struct contains information about possible ways from a junction.
- */
-struct TurnCandidate
-{
-  /*!
-   * angle is an angle of the turn in degrees. It means angle is 180 minus
-   * an angle between the current edge and the edge of the candidate. A counterclockwise rotation.
-   * The current edge is an edge which belongs the route and located before the junction.
-   * angle belongs to the range [-180; 180];
-   */
-  double angle;
-  /*!
-   * node is a possible node (a possible way) from the juction.
-   * May be NodeId for OSRM router or FeatureId::index for graph router.
-   */
-  uint32_t node;
-  /*!
-   * \brief highwayClass field for the road class caching. Because feature reading is a long
-   * function.
-   */
-  ftypes::HighwayClass highwayClass;
-
-  TurnCandidate(double a, uint32_t n, ftypes::HighwayClass c) : angle(a), node(n), highwayClass(c)
-  {
-  }
-};
-
-using TTurnCandidates = vector<TurnCandidate>;
-
-/*!
- * \brief The IRoutingResultGraph interface for the routing result. Uncouple router from the
- * annotation code that describes turns. See routers for detail implementations.
- */
-class IRoutingResultGraph
-{
-public:
-  virtual vector<LoadedPathSegment> const & GetSegments() const = 0;
-  virtual void GetPossibleTurns(NodeID node, m2::PointD const & ingoingPoint,
-                                m2::PointD const & junctionPoint,
-                                size_t & ingoingCount,
-                                TTurnCandidates & outgoingTurns) const = 0;
-  virtual double GetShortestPathLength() const = 0;
-  virtual m2::PointD const & GetStartPoint() const = 0;
-  virtual m2::PointD const & GetEndPoint() const = 0;
-
-  virtual ~IRoutingResultGraph() {}
-};
 
 /*!
  * \brief Compute turn and time estimation structs for the abstract route result.
